+++ apps/web/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @unkey/web
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [c1d1681]
+  - @unkey/ratelimit@0.1.1
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [ac4749c]
+- Updated dependencies [a89695f]
+  - @unkey/ratelimit@0.1.0
+
 ## 0.1.13
 
 ### Patch Changes