+++ packages/ratelimit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @unkey/ratelimit
 
+## 0.1.1
+
+### Patch Changes
+
+- c1d1681: Add async base config
+
+## 0.1.0
+
+### Minor Changes
+
+- ac4749c: Add async option
+
+### Patch Changes
+
+- a89695f: Update docs
+
 ## 0.0.3
 
 ### Patch Changes